Fixing double listing of file in SConscript on mac
Adding the MBGlob of #/include/ to a list of files using include/
ended up listing a file in both obj/include/ and ./include/, but
trying to install both to the same place.

There was no reason (as far as I know) to use #/include/, so I
removed it.

All headers should now be pulled only from obj/include/