BWCodegen: Temp fix for desktop builds
Prototype depends on a few birdwing header files from MBCoreUtils that are now code-generated, and birdwing code generation in MBCoreUtils depends on bw-scons-tools being a sibling directly so that the mustache_codegen SCons tool can be used.  The desktop toolchain doesn't pull in bw-scons-tools, so the code generation isn't run when building MBCoreUtils and thus the headers that Prototype depends on aren't created, causing desktop builds to fail.

as a temporary workaround:
moved mustache_codegen back into MBCoreUtils so bw-scons-tools isn't necessary to build

will revert this soon when a better solution is found