Different error messages for different sources
BW-2073
http://jira.makerbot.net/browse/BW-2073
Currently we only have one error code, not_connected, for which we
have different messaging for different error types.  The code gen
also allows for different error messages depending on the process
that is currently running and the step of the process on which the
error occurred, although there isn't currently any messaging in
place for this.

Since this code is also used by external clients, we still want to be
able the two error codes that are removed in this commit.  So I added
the ability to specify alternate error codes for each error.  The
alternate codes will not ever be produced by the machine, but will be
treated as the same error by any client which sees them.  Since the
older firmware which generates these codes still produces an error
source, the correct message will still be displayed.

I also modified the git hooks for this repo to ignore the cpp templates
in birdwing_codegen, since these are templates and are not expected to
be valid cpp.