Move errors.json to errors.py
BW-2073
http://jira.makerbot.net/browse/BW-2073

This allows us to properly separate the format for the error table from the
format for the code generated for that table.  It also lets us validate the
formatting of the error table so that we don't end up with difficult to track
down errors in our generated code.