Use new compile target stuff
Also, delete the Toolchain-Release compatibility because they have
switched to CMake, and it simplifies things.

This uses the new tools to make decisions about how to build. Nothing
builds differently, because nothing is compiled; it's all codegen. But
it still puts the outputs into the correct variant dirs so that behavior
is predictable across repos
BW-2587
http://jira.makerbot.net/browse/BW-2587