Don't crash on versions with strings
Firmware switched to using strings in their Version arrays. This patch
prevents those arrays from throwing exceptions, which prevented users
from uploading firmwares with verbose versions.

This is a bit of a hack since it just ignores verbose versions. I'm
doing it this way as a least-impact change to unblock uploading
verbose-versioned firmware. I'll add a PR to develop to actually parse
and handle verbose versioning.