Reconfigure tool mapping codegen
This was all to get pretty extruder type names in both c++ and
python. It has worked. I hate it anyway.

Types have keys, names, keystrings, and ids.
A _key_ is what is used to generate the names of enum members. It is
transformed into all-caps in cpp contexts.
A _keystring_ is what is parsed from metafiles. It's usually the same as
the key but it has to have a different name so it won't be transformed.
A _name_ is a pretty userfacing name for the type.
The _id_ is a number so pthon and cpp can translate between each other.

This shouldn't be a problem for code using this, though. The big feature
is the addition of a type_name_from_type method in python that lets you
get the user facing name.
BW-3020