Machine config: extruder profiles have "-" axes
When settings for extruders are stored in the "extruder_profiles"
subkey, any settings that apply on a per axis basis are now stored
with an axis name of "-" rather than "a". This makes it easy for
clients when applying extruder profile layers to rename those per-axis
settings to apply to the correct axis for dual extruder machines.