Don't specify too deep in machine config
Otherwise it breaks if the axis field doesn't have an "a" in it, which
is really not what we want.

This makes kaiten not crash when doing get_machine_config on sombrero,
and should work fine for birdwing as well. Note that this means the
sombrero machine config will look like
"extruder_profiles"/"mk14"/"steps_per_mm"/"-" still, since the
skeleton doesn't specify the content. This will make the time
estimator unhappy, we need a fix there to munge the multi-axis configs.