Lets just premptively add in a 13.3 and a 13.4
We are not going to be doing another birdwing release for a while, and I
generally don't want to have to do a release just to track tiny extruder
changes.  Also we don't really have any way to prevent people from ending
up with an extruder that is too new for their current firmware to handle
and as far as I know the messaging for this would just tell them that their
extruder is broken.