Better syntactic sugar for SafeJson
BW-4289
http://jira.makerbot.net/browse/BW-4289

The orginal set of SafeJson functions was an interesting concept, but all of
the code using these functions is an unreadable mess and this really seems to
be inherent in the architecture of these functions.  So I started from how I
would want the code to look for actually using these functions and then wrote
this implementation based on that.