Use fundamental integer types for overloading
BW-4289
http://jira.makerbot.net/browse/BW-4289

The old method didn't actually have any overloads for size_t on osx, because
uint32_t is an unsigned int, uint64_t is an unsigned long long, and size_t is
an unsigned long.  But we couldn't just add an overload for size_t because in
64 bit linux size_t and uint64_t are both unsigned long and that would create
an ambiguous overload.  Using the base distinct types instead of aliases when
defining our overloads is the only way to write this that is guaranteed to
work according to the C++ standard (so 50/50 whether it works on windows -- in
fact I left out wide character types because I didn't want to push my luck).