add dragon fatal error. remove zero cross error.
Dragon has a lot of errors and they are all logged to the dragon log
file now. The most common errors are already passed as commonly
available errors. The rest will just fall under the fatal category and
then these can just be looked up in the log file. The common errors
would be standard things like a fan or temperature sonsor error.