Move boost logging setup to a common location
BW-5632
http://makerbot.atlassian.net/browse/BW-5632

This takes the boost logging setup that we copied across almost a dozen
repositories and places a single version here.  The header and source
files are now both in the header, with all functions marked as inline
to prevent ODR errors.  Static variables are also transitioned to local
static variables, which is supposedly necessary to ensure that we don't
create separate copies of these variables for each translation unit, but
I did run some tests and it does _seem_ like static variables are not
getting duplicated across translation units.

One nuisance about moving all logging to a common location is that we
were very much relying on each repository that set up logging to use a
completely unique set of names for every variable/function/class defined
for the general log in order to actually create distinct log sources
with distinct channel names.  We could try to use the channel name to
generate distinct names via macros but then we wouldn't be able to use
libstepper/pru as a channel name anymore.  The only thing we are
currently using distinct log sources within the general logs is to put
the channel name in the log message, so we just use the logging macro
to put the channel name into the log message instead.

TODO: LOGGING_CHANNEL is now a misleading name...