Move boost logging setup to a common location
BW-5632
http://makerbot.atlassian.net/browse/BW-5632

This takes the boost logging setup that we copied across almost a dozen
repositories and places a single version here.  The header and source
files are now both in the header, with all functions marked as inline
to prevent ODR errors.  Static variables are also transitioned to local
static variables, which is supposedly necessary to ensure that we don't
create separate copies of these variables for each translation unit, but
I did run some tests and it does _seem_ like static variables are not
getting duplicated across translation units.

One nuisance about moving all logging to a common location is that we
were very much relying on each repository that set up logging to use a
completely unique set of names for every variable/function/class defined
for the general log in order to actually create distinct log sources
with distinct channel names.  We could try to use the channel name to
generate distinct names via macros but then we wouldn't be able to use
libstepper/pru as a channel name anymore.  The only thing we are
currently using distinct log sources within the general logs is to put
the channel name in the log message, so we just use the logging macro
to put this name into the log message instead.  Since this is now fully
unrelated to the actual log channels (which we still use), we rename
this to be a "context" instead of a "channel"

The __FILENAME__ macro from the old logging setup has been replaced by
passing the __FILE__ macro through a constexpr function, which has the
advantage of getting the filename right even if we log from a header
file (which we now do here in fact) and not needing build system support
(though before this solution I did have the build system support down to
a single utility function that would automatically add the macro to
every single source file for a given target).  The disadvantages are
that I couldn't find any C++11 way to force compile time evaluation of
the function without breaking the macro (I manually checked that the
evaluation does not happen at runtime but the compiler would be allowed
to evaluate at runtime and still be compliant) and that the full file
path is still stored in rodata (the compiler can't verify that we don't
try to read backwards from the start of the substring).