Use a custom collector for logs
BW-5632
http://makerbot.atlassian.net/browse/BW-5632

A major issue that we have with the logs is that we apply a single size
cap for both telemetry and general logs.  General logs hold less data but
generally much more useful data which we would want to hold on to for a
long period of time, and telemetry data is quite voluminous but usually
not as valuable.  By combining both under a single collector, whenever
telemetry is enabled we generally completely wipe out all general logs
except the most recent, so as a result we avoid enabling telemetry.

A second issue that we sometimes have with logs is that we do sometimes
delete newer logs before clearing out older logs, because printers don't
always have the correct time and the default collector only considers
file timestamps to determine which entries to delete.

We could get around the first issue by merely keeping telemetry logs in
a separate folder.  This would be a bit of a pain to manage to make sure
that old telemetry logs get purged on firmware update, but would likely
be a preferable solution if this was the only problem.  But since a
custom collector is the only way to address the second issue, we use a
custom collector to address both problems.

Some less than ideal syntax is used here to continue to support older
boost versions for host compilation (frankly I am surprised that it is
even possible to support these older boost versions).  We certainly
could just #ifdef out the custom collector code for old boost versions
since we only really need this to work on the firmware (and are we
really ever going to write unit tests for this)?