Change timestamp format for logs
BW-5632
http://makerbot.atlassian.net/browse/BW-5632

Every timestamp is now both an ISO 8601 timestamp with timezone followed by a
decimal number of seconds since boot up, both of which are specified with
microsecond precision.  (This matches the kaiten format exactly except that
kaiten includes a colon in the time zone offset, which is optional in ISO
8601.)

This is implemeted by adding a new boost logging attribute that stores both
timestamps when constructed and has an ostream overload to provide the format
in the logs.  The actual code for retrieving and formatting timestamps just
uses c++11 because using boost datetime was a huge mess and still couldn't
get the timezone right.  (The C++11 implementation does use std::localtime
which is not re-entrant, but we really just need the date to _mostly_ be
right and I also doubt it is used anywhere else in the entire system outside
of this file so in practice it will still always be right.)

This also removes the "standard" boost log attributes because we now don't
actually use any of them.