Add boot id to the start of all logs
BW-5632
http://makerbot.atlassian.net/browse/BW-5632

This single log statement serves both to identify the boot id for each
run of monotonic timestamps (so that within each boot all events can be
precisely aligned) and to identify the startup of whatever daemon was
logging, so that we won't get confused by restarted daemons or daemons
that just fail to start up.  For the boot id we just read from the path
where it exists in firmware, but just fall back to an empty string if
anything fails, as we generally expect when host compiling.

The logic for telemetry is a bit complex, because we don't want to write
any telemetry entries without this initial log statement, but we also
don't want to write this header if telemetry never gets enabled or we
will have a million ui and accessories telemetry logs that literally
only have this header.