Workaround for a weird pystache bug
Our old version of pystash did not support python 3, so we checked in a
newer version of it that does support python 3.  However it appears that
this newer version has a minor bug that is only occurring on our windows
builders, where if we use a {{#value}}...{{/value}} block to include a
block of code only if the value key is present and the value key is a
string, the string is actually treated as a list of characters and the
block of code is repeated once for every character in the string.

The right thing to do would probably be to track down the pystache bug
and fix it, but it only occurs on windows so I would have to debug it on
windows and I _really_ don't want to do that.  As it happens, we only use
this construct here to select between two different naming conventions in
some generated code, and the context only ever contains the naming key
for exactly one of the two conventions, so we should actually just be
able to avoid all use of the broken construct.